2018-12-04 bug2 创建许多的range，但是只应该有一个

+++ tools/db_bench_tool.cc
@@ -4143,7 +4143,7 @@ private:
                 int64_t rand_num = key_gens[id]->Next();
                 GenerateKeyFromInt(rand_num, FLAGS_num, &key);
                 DBG_PRINT("gen key[%ld]to[%s]/FLAGS_num[%ld]",rand_num, std::string(key.data(), key.size()).c_str(), FLAGS_num);
-                for(int i = 0; i < key.size(); i++){
+                for(size_t i = 0; i < key.size(); i++){
                     printf("%d ", key.data()[i]);
                 }
                 printf("\n");
